simplify set, remove separate set utility files
