+++ packages/manager/src/utilities/formikErrorUtils.test.ts
@@ -2,6 +2,7 @@ import {
   getFormikErrorsFromAPIErrors,
   handleAPIErrors,
   handleVPCAndSubnetErrors,
+  set,
 } from './formikErrorUtils';
 
 const errorWithoutField = [{ reason: 'Internal server error' }];
@@ -199,5 +200,197 @@ describe('getFormikErrorsFromAPIErrors', () => {
     for (const { apiErrors, expected } of testCases) {
       expect(getFormikErrorsFromAPIErrors(apiErrors)).toEqual(expected);
     }
+  });
+});
+
+describe('Tests for set', () => {
+  it("returns the passed in 'object' as is if it's not actually a (non array) object", () => {
+    expect(set([], 'path not needed', 3)).toEqual([]);
+  });
+
+  describe('Correctly setting the value at the given path', () => {
+    it('sets the value for a simple path', () => {
+      const object = {};
+      let settedObject = set(object, 'test', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({ test: 1 });
+
+      settedObject = set(object, 'test2', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({ test: 1, test2: 1 });
+    });
+
+    it('sets the value for complex string paths (without indexes)', () => {
+      const object = {};
+
+      set(object, 'a.b.c', 'c');
+      expect(object).toEqual({ a: { b: { c: 'c' } } });
+
+      set(object, 'a.b.d', 'd');
+      expect(object).toEqual({
+        a: { b: { c: 'c', d: 'd' } },
+      });
+
+      set(object, 'e[f][g]', 'g');
+      expect(object).toEqual({
+        a: { b: { c: 'c', d: 'd' } },
+        e: { f: { g: 'g' } },
+      });
+    });
+
+    it('sets the value for complex string paths (with indexes)', () => {
+      const object = {};
+
+      set(object, 'a.b.1', 'b1');
+      expect(object).toEqual({ a: { b: [undefined, 'b1'] } });
+      set(object, 'a.b[0]', 5);
+      expect(object).toEqual({ a: { b: [5, 'b1'] } });
+
+      set(object, 'a.b.2', 'b2');
+      expect(object).toEqual({
+        a: { b: [5, 'b1', 'b2'] },
+      });
+
+      set(object, 'a.b[3].c', 'c');
+      expect(object).toEqual({
+        a: { b: [5, 'b1', 'b2', { c: 'c' }] },
+      });
+    });
+
+    it('only considers 0 or positive integers for setting array values', () => {
+      const object = {};
+
+      expect(set(object, 'test[-01].test1', 'test')).toEqual({
+        test: { '-01': { test1: 'test' } },
+      });
+      expect(set(object, 'test[-01][-02]', 'test2')).toEqual({
+        test: { '-01': { '-02': 'test2', test1: 'test' } },
+      });
+      expect(set(object, 'test[   02]', 'test3')).toEqual({
+        test: {
+          '   02': 'test3',
+          '-01': { '-02': 'test2', test1: 'test' },
+        },
+      });
+      expect(set(object, 'test[0 0]', 'test4')).toEqual({
+        test: {
+          '   02': 'test3',
+          '-01': { '-02': 'test2', test1: 'test' },
+          '0 0': 'test4',
+        },
+      });
+    });
+
+    it('considers numbers as keys if they are not followed by another number or if there is an already existing object', () => {
+      const object = {};
+      set(object, '1', 'test');
+      expect(object).toEqual({ 1: 'test' });
+
+      set(object, '2', '2');
+      expect(object).toEqual({ 1: 'test', 2: '2' });
+
+      expect(set({ test: { test1: 'test' } }, 'test[1]', 'test2')).toEqual({
+        test: { '1': 'test2', test1: 'test' },
+      });
+    });
+
+    it('treats numbers as array indexes if they precede some previous key (if they are valid indexes)', () => {
+      const obj1 = set({}, '1[1]', 'test');
+      expect(obj1).toEqual({ 1: [undefined, 'test'] });
+
+      const obj2 = set({}, '1.2', 'test');
+      expect(obj2).toEqual({ 1: [undefined, undefined, 'test'] });
+    });
+
+    it('can replace the value at an already existing key', () => {
+      const alreadyExisting = { test: 'test' };
+      expect(set(alreadyExisting, 'test', 'changed')).toEqual({
+        test: 'changed',
+      });
+      expect(set(alreadyExisting, 'test[test2][test3]', 'changed x4')).toEqual({
+        test: { test2: { test3: 'changed x4' } },
+      });
+    });
+
+    it('sets the value for nonstandard paths', () => {
+      expect(set({}, 'test.[.test]', 'testing 2')).toEqual({
+        test: { test: 'testing 2' },
+      });
+      expect(set({}, 'test.[te[st]', 'testing 3')).toEqual({
+        test: { te: { st: 'testing 3' } },
+      });
+      expect(set({}, 'test.]test', 'testing 4')).toEqual({
+        test: { test: 'testing 4' },
+      });
+    });
+  });
+
+  describe('Ensuring safety against prototype pollution and that the passed in and returned object are the same', () => {
+    it('protects against the given string path matching a prototype pollution key', () => {
+      const object = {};
+      // __proto__
+      let settedObject = set(object, '__proto__', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({});
+
+      // constructor
+      settedObject = set(object, 'constructor', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({});
+
+      // prototype
+      settedObject = set(object, 'prototype', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({});
+    });
+
+    it('protects against the given string path containing prototype pollution keys that are separated by path delimiters', () => {
+      const object = {};
+      // prototype pollution key separated by .
+      let settedObject = set(object, 'test.__proto__.test', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({});
+
+      settedObject = set(object, 'test.constructor.test', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({});
+
+      settedObject = set(object, 'test.prototype.test', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({});
+
+      // prototype pollution key separated by []
+      settedObject = set(object, 'test.test[__proto__]', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({});
+
+      settedObject = set(object, 'test.test[constructor]', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({});
+
+      settedObject = set(object, 'test.test[prototype]', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({});
+    });
+
+    it('is not considered prototype pollution if the string paths have a key not separated by delimiters', () => {
+      const object = {};
+      // prototype pollution key separated by .
+      let settedObject = set(object, 'test__proto__test', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({ test__proto__test: 1 });
+
+      settedObject = set(object, 'constructortest', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({ constructortest: 1, test__proto__test: 1 });
+
+      settedObject = set(object, 'testprototype', 1);
+      expect(object).toBe(settedObject);
+      expect(object).toEqual({
+        constructortest: 1,
+        test__proto__test: 1,
+        testprototype: 1,
+      });
+    });
   });
 });
