--- packages/manager/src/utilities/set.ts
@@ -1,111 +1,0 @@
-type PropertyName = number | string | symbol;
-type PropertyPath = PropertyName | readonly PropertyName[];
-
-/**
- * Helper to set the given value at the given path of the given object.
- * This util serves as a replacement for `set` from lodash, with additional checks
- * to protect against prototype pollution. It is not (yet) a perfect replacement (see
- * skipped test examples in set.test.ts), but does handle current CM use cases.
- *
- * @param object — The object to modify.
- * @param path — The path of the property to set.
- * @param value — The value to set.
- * @return — Returns object.
- */
-export function set<T extends object>(
-  object: T,
-  path: PropertyPath,
-  value: any
-): T {
-  // ensure that object is actually a non-array object
-  if (!object || Array.isArray(object) || typeof object !== 'object') {
-    return object;
-  }
-
-  // if the path is not an array, convert it to an array format
-  const updatedPath = determinePath(path);
-
-  if (
-    // ensure that the path will not lead to prototype pollution issues
-    // and that there is actually a path to set
-    !isPrototypePollutionSafe(updatedPath) ||
-    updatedPath.length === 0
-  ) {
-    return object;
-  }
-
-  let updatingObject: any = object;
-  // if we haven't reached the last key, traverse through object and
-  for (let i = 0; i < updatedPath.length - 1; i++) {
-    const key = updatedPath[i];
-    // set object[key] to be [] or {} depending on the next key
-    if (!updatingObject[key] || typeof updatingObject[key] !== 'object') {
-      const nextKey = updatedPath[i + 1];
-      updatingObject[key] = isValidIndex(nextKey) ? [] : {};
-    }
-
-    updatingObject = updatingObject[key];
-  }
-
-  // set value after reaching end of path
-  updatingObject[updatedPath[updatedPath.length - 1]] = value;
-
-  return object;
-}
-
-/**
- * -------------------------------------------------------------------------
- * Helper functions for set
- * -------------------------------------------------------------------------
- */
-
-/**
- * Determines the path to set some value at, converting any string paths
- * into array format.
- *
- * @param path - The path to check
- * @returns - The path to set a value at, in array format
- */
-export const determinePath = (path: PropertyPath): PropertyName[] => {
-  if (Array.isArray(path)) {
-    return path;
-  }
-
-  return typeof path === 'string'
-    ? path.toString().match(/[^.[\]]+/g) ?? ['']
-    : ([path] as PropertyName[]);
-};
-
-/**
- * Ensures a path cannot lead to a prototype pollution issue.
- *
- * @param path - The path to check
- * @return - boolean depending on whether the path is safe or not
- */
-export const isPrototypePollutionSafe = (path: PropertyName[]): boolean => {
-  return path.reduce((safeSoFar, val) => {
-    const isCurKeySafe =
-      typeof val === 'string'
-        ? val !== '__proto__' && val !== 'prototype' && val !== 'constructor'
-        : true;
-    return safeSoFar && isCurKeySafe;
-  }, true);
-};
-
-/**
- * Determines if the given value can be considered a valid index for an array/string.
- * For example, 0, 1, 2 are all valid indexes.
- * -1, 01, -01, 00, '1 1' are not.
- *
- * @param value - The value to check
- *
- * @returns - Returns boolean depending on if value is a valid index or not
- */
-export const isValidIndex = (value: PropertyName) => {
-  const convertedValue = value.toString();
-
-  return (
-    /^\d+$/.test(convertedValue) && // must be all digits
-    (convertedValue.length < 2 || !convertedValue.startsWith('0')) // must not start with 0 (unless it is 0 only)
-  );
-};
